Fixed onPhotoTap() not called when zoom disabled
There was an overly broad check for mZoomEnabled in onTouch() that
prevented zooming but also prevented the onPhotoTap() listeners from
getting their events. I wanted to just skip mScaleDragDetector all
together when zoom was disabled, but it seems to be required for
singleTap detection. Might be better if mGestureDetector handled both
single and double tap detection.