Fix for regression on issue #72.
If the PhotoView gesture detector handles scaling or dragging the PhotoView, we will continue to block the parent view from receiving onTouch events. If the gesture detector did not scale or drag the PhotoView at all, we allow the parent to properly receive all touch events.

Tested on 4.x and 5.x in an Activity with both a NavigationDrawer and a ViewPager. Both components work as designed along with scaling/panning a PhotoView.