Translate image during scale
This commit changes the behaviour of the scale action. Until now when the user scaled the image and the focus point has changed as the user moved his fingers around the screen, we did not translate the image.
Now we translate the image as well allowing the user to zoom and pan the image simultaneously.